chore: Break lyric overlay code to its own file
